+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix(asea): failure in network association stack with security groups associated with RAM shared subnets/vpc
 - fix(asea): fixed mapping upload logic to use lza enabled regions and accounts
 - fix(networking): fixed hosted zones for sagemaker vpc endpoints
+- fix(service quota): fix region logic that provisioned service limit for home region even if not specified
 
 ### Changed